Send client handshake when server binds
At the moment, it's possible for the server to miss the client's `hs`
handshake message. Since it never gets the handshake, it never sends one
back, and the client remains uninitialised.

This can happen when performing asynchronous, long-running tasks between
opening the websocket, and calling `backend.listen` to bind the socket
to ShareDB.

This change fixes this issue by having the client send the handshake
message twice:

 - Once when the `Connection` is initialised. If the server is ready,
   but the client connected late, we'll still handshake.
 - Once when the client receives the server's `init` message. If the
   server bound to the socket late, and missed the first handshake
   message, this prompts us to send it again.